providers get own client component
