+++ prettier.config.js
@@ -1,7 +1,7 @@
 // prettier.config.js
 module.exports = {
-	singleQuote: false,
-	tabWidth: 2,
-	semi: true,
-	plugins: [require("prettier-plugin-tailwindcss")],
+  singleQuote: false,
+  tabWidth: 2,
+  semi: true,
+  plugins: ["prettier-plugin-tailwindcss"],
 };
